evaluator-gcp-dataproc/requirements.txt
google-cloud-dataproc
google-cloud-storage
